requirements.txt
Django>=3.0.1,<3.1.0
djangorestframework>=3.11.0,<3.12.0
South>=1.0.2,<1.1.0
psycopg2>=2.8.4,<2.9.0
Pillow>=7.0.0,<7.1.0